# src/main/resources/api.yml
openapi: "3.0.0"
info:
  title: Tiny ledger API
  description: Tiny ledger API.
  version: 0.0.1
paths:
  /account:
    get:
      summary: List all accounts
      operationId: listAccounts
      tags:
        - accounts
      responses:
        200:
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      summary: Create new account
      operationId: createAccount
      tags:
        - accounts
      responses:
        200:
          description: Account reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountReference"
  /account/{accountId}:
    get:
      summary: Get specific account
      operationId: getAccount
      tags:
        - accounts
      parameters:
        - name: accountId
          in: path
          description: "ID of the account to return"
          required: true
          schema:
            type: string
      responses:
        200:
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        404:
          description: Account not found.
  /account/{accountId}/movement:
    post:
      summary: Create a money movement on an account.
      operationId: createMovement
      tags:
        - accounts
      parameters:
        - name: accountId
          in: path
          description: "ID of the account"
          required: true
          schema:
            type: string
        - name: amount
          in: query
          description: Value for the money movement
          schema:
            type: integer
        - name: type
          in: query
          description: Type of the money movement
          schema:
            type: string
            enum:
              - DEPOSIT
              - WITHDRAWAL
      responses:
        200:
          description: Created money movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoneyMovement"
        400:
          description: Validation error.
        404:
          description: Account not found.
    get:
      summary: List account movements.
      operationId: listMovements
      tags:
        - accounts
      parameters:
        - name: accountId
          in: path
          description: "ID of the account"
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoneyMovement"
        404:
          description: Account not found.


components:
  schemas:
    AccountReference:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    Account:
      type: object
      required:
        - id
        - balance
      properties:
        id:
          type: string
        balance:
          type: integer
    MoneyMovement:
      type: object
      required:
        - type
        - value
      properties:
        id:
          type: string
        amount:
          type: integer
        type:
          type: string
          enum:
            - DEPOSIT
            - WITHDRAWAL
        createdAt:
          type: string
          format: date-time
        effectedAt:
          type: string
          format: date-time